refactor: created library functions
