+++ packages/qti-transformers/src/item/stylesheets.ts
@@ -1,0 +1,4 @@
+export function stripStyleSheets(xmlFragment: DocumentFragment) {
+  // remove qti-stylesheet tag
+  xmlFragment.querySelectorAll('qti-stylesheet').forEach(stylesheet => stylesheet.remove());
+}
