+++ packages/qti-transformers/src/test/items.ts
@@ -1,0 +1,10 @@
+export function itemsFromTest(xmlFragment: DocumentFragment) {
+  const items: { identifier: string; href: string; category: string }[] = [];
+  xmlFragment.querySelectorAll('qti-assessment-item-ref').forEach(el => {
+    const identifier = el.getAttribute('identifier');
+    const href = el.getAttribute('href');
+    const category = el.getAttribute('category');
+    items.push({ identifier, href, category });
+  });
+  return items;
+}
